+++ actionpack/CHANGELOG.md
@@ -1,3 +1,25 @@
+*   Accept render options and block in `render` calls made with `:renderable`
+
+    ```ruby
+    class Greeting
+      def render_in(view_context, **options, &block)
+        if block
+          view_context.render html: block.call
+        else
+          view_context.render inline: <<~ERB.strip, **options
+            Hello, <%= local_assigns.fetch(:name, "World") %>!
+          ERB
+      end
+    end
+
+    ApplicationController.render(Greeting.new)                                        # => "Hello, World!"
+    ApplicationController.render(Greeting.new) { "Hello, Block!" }                    # => "Hello, Block!"
+    ApplicationController.render(renderable: Greeting.new)                            # => "Hello, World!"
+    ApplicationController.render(renderable: Greeting.new, locals: { name: "Local" }) # => "Hello, Local!"
+    ```
+
+    *Sean Doyle*
+
 *   Request Forgery takes relative paths into account.
 
     *Stefan Wienert*